Fixed for 64-bit architecture

+++ System Services/Utilities/SSNetworkInfo.m
@@ -15,34 +15,34 @@
 // Get Current IP Address
 + (NSString *)currentIPAddress {
     // Get the current IP Address
-    
+
     // Check which interface is currently in use
     if ([self connectedToWiFi]) {
         // WiFi is in use
-        
+
         // Get the WiFi IP Address
         NSString *WiFiAddress = [self wiFiIPAddress];
-        
+
         // Check that you get something back
         if (WiFiAddress == nil || WiFiAddress.length <= 0) {
             // Error, no address found
             return nil;
         }
-        
+
         // Return Wifi address
         return WiFiAddress;
     } else if ([self connectedToCellNetwork]) {
         // Cell Network is in use
-        
+
         // Get the Cell IP Address
         NSString *CellAddress = [self cellIPAddress];
-        
+
         // Check that you get something back
         if (CellAddress == nil || CellAddress.length <= 0) {
             // Error, no address found
             return nil;
         }
-        
+
         // Return Cell address
         return CellAddress;
     } else {
@@ -54,34 +54,34 @@
 // Get Current MAC Address
 + (NSString *)currentMACAddress {
     // Get the current interface MAC Address
-    
+
     // Check which interface is currently in use
     if ([self connectedToWiFi]) {
         // WiFi is in use
-        
+
         // Get the WiFi MAC Address
         NSString *WiFiAddress = [self wiFiMACAddress];
-        
+
         // Check that you get something back
         if (WiFiAddress == nil || WiFiAddress.length <= 0) {
             // Error, no address found
             return nil;
         }
-        
+
         // Return Wifi address
         return WiFiAddress;
     } else if ([self connectedToCellNetwork]) {
         // Cell Network is in use
-        
+
         // Get the Cell MAC Address
         NSString *CellAddress = [self cellMACAddress];
-        
+
         // Check that you get something back
         if (CellAddress == nil || CellAddress.length <= 0) {
             // Error, no address found
             return nil;
         }
-        
+
         // Return Cell address
         return CellAddress;
     } else {
@@ -98,7 +98,7 @@
             // Not connected to anything, return nil
             return nil;
         }
-        
+
         // Get the external IP Address based on dynsns.org
         NSError *error = nil;
         NSString *theIpHtml = [NSString stringWithContentsOfURL:[NSURL URLWithString:@"http://www.dyndns.org/cgi-bin/check_ip.cgi"]
@@ -110,17 +110,17 @@
             NSString *externalIP;
             NSScanner *theScanner;
             NSString *text = nil;
-            
+
             theScanner = [NSScanner scannerWithString:theIpHtml];
-            
+
             while ([theScanner isAtEnd] == NO) {
-                
+
                 // find start of tag
                 [theScanner scanUpToString:@"<" intoString:NULL] ;
-                
+
                 // find end of tag
                 [theScanner scanUpToString:@">" intoString:&text] ;
-                
+
                 // replace the found tag with a space
                 //(you can filter multi-spaces out later if you wish)
                 theIpHtml = [theIpHtml stringByReplacingOccurrencesOfString:
@@ -128,16 +128,16 @@
                                                                  withString:@" "] ;
                 ipItemsArray = [theIpHtml  componentsSeparatedByString:@" "];
                 an_Integer = [ipItemsArray indexOfObject:@"Address:"];
-                
+
                 externalIP =[ipItemsArray objectAtIndex:++an_Integer];
             }
-            
+
             // Check that you get something back
             if (externalIP == nil || externalIP.length <= 0) {
                 // Error, no address found
                 return nil;
             }
-            
+
             // Return External IP
             return externalIP;
         } else {
@@ -162,13 +162,13 @@
         struct ifaddrs *Temp;
         struct sockaddr_in *s4;
         char buf[64];
-        
+
         // If it's 0, then it's good
         if (!getifaddrs(&Interfaces))
         {
             // Loop through the list of interfaces
             Temp = Interfaces;
-            
+
             // Run through it while it's still available
             while(Temp != NULL)
             {
@@ -179,7 +179,7 @@
                     if([[NSString stringWithUTF8String:Temp->ifa_name] isEqualToString:@"pdp_ip0"])
                     {
                         s4 = (struct sockaddr_in *)Temp->ifa_addr;
-                        
+
                         if (inet_ntop(Temp->ifa_addr->sa_family, (void *)&(s4->sin_addr), buf, sizeof(buf)) == NULL) {
                             // Failed to find it
                             IPAddress = nil;
@@ -189,21 +189,21 @@
                         }
                     }
                 }
-                
+
                 // Set the temp value to the next interface
                 Temp = Temp->ifa_next;
             }
         }
-        
+
         // Free the memory of the interfaces
         freeifaddrs(Interfaces);
-        
+
         // Check to make sure it's not empty
         if (IPAddress == nil || IPAddress.length <= 0) {
             // Empty, return not found
             return nil;
         }
-        
+
         // Return the IP Address of the WiFi
         return IPAddress;
     }
@@ -224,14 +224,14 @@
         unsigned char       macAddress[6];
         struct if_msghdr    *interfaceMsgStruct;
         struct sockaddr_dl  *socketStruct;
-        
+
         // Setup the management Information Base (mib)
         mgmtInfoBase[0] = CTL_NET;        // Request network subsystem
         mgmtInfoBase[1] = AF_ROUTE;       // Routing table info
         mgmtInfoBase[2] = 0;
         mgmtInfoBase[3] = AF_LINK;        // Request link layer information
         mgmtInfoBase[4] = NET_RT_IFLIST;  // Request all configured interfaces
-        
+
         // With all configured interfaces requested, get handle index
         if ((mgmtInfoBase[5] = if_nametoindex([@"pdp_ip0" UTF8String])) == 0)
             // Error, Name to index failure
@@ -257,33 +257,33 @@
                 }
             }
         }
-        
+
         // Map msgbuffer to interface message structure
         interfaceMsgStruct = (struct if_msghdr *) msgBuffer;
-        
+
         // Map to link-level socket structure
         socketStruct = (struct sockaddr_dl *) (interfaceMsgStruct + 1);
-        
+
         // Copy link layer address data in socket structure to an array
         memcpy(&macAddress, socketStruct->sdl_data + socketStruct->sdl_nlen, 6);
-        
+
         // Read from char array into a string object, into traditional Mac address format
         NSString *macAddressString = [NSString stringWithFormat:@"%02X:%02X:%02X:%02X:%02X:%02X",
                                       macAddress[0], macAddress[1], macAddress[2],
                                       macAddress[3], macAddress[4], macAddress[5]];
-        
+
         // Release the buffer memory
         free(msgBuffer);
-        
+
         // Make a new string from the macAddressString
         NSString *deviceID = macAddressString;
-        
+
         // If the device ID comes back empty
         if (deviceID == (id)[NSNull null] || deviceID.length <= 0) {
             // Return that the MAC address was not found
             return nil;
         }
-        
+
         // Return Successful
         return deviceID;
     }
@@ -303,13 +303,13 @@
         strncpy(afr.ifr_name, [@"pdp_ip0" UTF8String], IFNAMSIZ-1);
         // Open a socket
         int afd = socket(AF_INET, SOCK_DGRAM, 0);
-        
+
         // Check the socket
         if (afd == -1) {
             // Error, socket failed to open
             return nil;
         }
-        
+
         // Check the netmask output
         if (ioctl(afd, SIOCGIFNETMASK, &afr) == -1) {
             // Error, netmask wasn't found
@@ -318,22 +318,22 @@
             // Return error
             return nil;
         }
-        
+
         // Close the socket
         close(afd);
-        
+
         // Create a char for the netmask
         char *netstring = inet_ntoa(((struct sockaddr_in *)&afr.ifr_addr)->sin_addr);
-        
+
         // Create a string for the netmask
         NSString *Netmask = [NSString stringWithUTF8String:netstring];
-        
+
         // Check to make sure it's not nil
         if (Netmask == nil || Netmask.length <= 0) {
             // Error, netmask not found
             return nil;
         }
-        
+
         // Return successful
         return Netmask;
     }
@@ -350,7 +350,7 @@
         // Set up strings for the IP and Netmask
         NSString *IPAddress = [self cellIPAddress];
         NSString *NMAddress = [self cellNetmaskAddress];
-        
+
         // Check to make sure they aren't nil
         if (IPAddress == nil || IPAddress.length <= 0) {
             // Error, IP Address can't be nil
@@ -360,41 +360,41 @@
             // Error, NM Address can't be nil
             return nil;
         }
-        
+
         // Check the formatting of the IP and NM Addresses
         NSArray *IPCheck = [IPAddress componentsSeparatedByString:@"."];
         NSArray *NMCheck = [NMAddress componentsSeparatedByString:@"."];
-        
+
         // Make sure the IP and NM Addresses are correct
         if (IPCheck.count != 4 || NMCheck.count != 4) {
             // Incorrect IP Addresses
             return nil;
         }
-        
+
         // Set up the variables
         NSUInteger IP = 0;
         NSUInteger NM = 0;
         NSUInteger CS = 24;
-        
+
         // Make the address based on the other addresses
         for (NSUInteger i = 0; i < 4; i++, CS -= 8) {
             IP |= [[IPCheck objectAtIndex:i] intValue] << CS;
             NM |= [[NMCheck objectAtIndex:i] intValue] << CS;
         }
-        
+
         // Set it equal to the formatted raw addresses
         NSUInteger BA = ~NM | IP;
-        
+
         // Make a string for the address
-        NSString *BroadcastAddress = [NSString stringWithFormat:@"%d.%d.%d.%d", (BA & 0xFF000000) >> 24,
-                                      (BA & 0x00FF0000) >> 16, (BA & 0x0000FF00) >> 8, BA & 0x000000FF];
-        
+        NSString *BroadcastAddress = [NSString stringWithFormat:@"%lu.%lu.%lu.%lu", (unsigned long)(BA & 0xFF000000) >> 24,
+				(unsigned long)(BA & 0x00FF0000) >> 16, (unsigned long)(BA & 0x0000FF00) >> 8, (unsigned long)BA & 0x000000FF];
+
         // Check to make sure the string is valid
         if (BroadcastAddress == nil || BroadcastAddress.length <= 0) {
             // Error, no address
             return nil;
         }
-        
+
         // Return Successful
         return BroadcastAddress;
     }
@@ -415,16 +415,16 @@
         struct ifaddrs *Temp;
         // Set up int for success or fail
         int Status = 0;
-        
+
         // Get all the network interfaces
         Status = getifaddrs(&Interfaces);
-        
+
         // If it's 0, then it's good
         if (Status == 0)
         {
             // Loop through the list of interfaces
             Temp = Interfaces;
-            
+
             // Run through it while it's still available
             while(Temp != NULL)
             {
@@ -438,21 +438,21 @@
                         IPAddress = [NSString stringWithUTF8String:inet_ntoa(((struct sockaddr_in *)Temp->ifa_addr)->sin_addr)];
                     }
                 }
-                
+
                 // Set the temp value to the next interface
                 Temp = Temp->ifa_next;
             }
         }
-        
+
         // Free the memory of the interfaces
         freeifaddrs(Interfaces);
-        
+
         // Check to make sure it's not empty
         if (IPAddress == nil || IPAddress.length <= 0) {
             // Empty, return not found
             return nil;
         }
-        
+
         // Return the IP Address of the WiFi
         return IPAddress;
     }
@@ -473,14 +473,14 @@
         unsigned char       macAddress[6];
         struct if_msghdr    *interfaceMsgStruct;
         struct sockaddr_dl  *socketStruct;
-        
+
         // Setup the management Information Base (mib)
         mgmtInfoBase[0] = CTL_NET;        // Request network subsystem
         mgmtInfoBase[1] = AF_ROUTE;       // Routing table info
         mgmtInfoBase[2] = 0;
         mgmtInfoBase[3] = AF_LINK;        // Request link layer information
         mgmtInfoBase[4] = NET_RT_IFLIST;  // Request all configured interfaces
-        
+
         // With all configured interfaces requested, get handle index
         if ((mgmtInfoBase[5] = if_nametoindex([@"en0" UTF8String])) == 0)
             // Error, Name to index failure
@@ -506,33 +506,33 @@
                 }
             }
         }
-        
+
         // Map msgbuffer to interface message structure
         interfaceMsgStruct = (struct if_msghdr *) msgBuffer;
-        
+
         // Map to link-level socket structure
         socketStruct = (struct sockaddr_dl *) (interfaceMsgStruct + 1);
-        
+
         // Copy link layer address data in socket structure to an array
         memcpy(&macAddress, socketStruct->sdl_data + socketStruct->sdl_nlen, 6);
-        
+
         // Read from char array into a string object, into traditional Mac address format
         NSString *macAddressString = [NSString stringWithFormat:@"%02X:%02X:%02X:%02X:%02X:%02X",
                                       macAddress[0], macAddress[1], macAddress[2],
                                       macAddress[3], macAddress[4], macAddress[5]];
-        
+
         // Release the buffer memory
         free(msgBuffer);
-        
+
         // Make a new string from the macAddressString
         NSString *deviceID = macAddressString;
-        
+
         // If the device ID comes back empty
         if (deviceID == (id)[NSNull null] || deviceID.length <= 0) {
             // Return that the MAC address was not found
             return nil;
         }
-        
+
         // Return Successful
         return deviceID;
     }
@@ -552,13 +552,13 @@
         strncpy(afr.ifr_name, [@"en0" UTF8String], IFNAMSIZ-1);
         // Open a socket
         int afd = socket(AF_INET, SOCK_DGRAM, 0);
-        
+
         // Check the socket
         if (afd == -1) {
             // Error, socket failed to open
             return nil;
         }
-        
+
         // Check the netmask output
         if (ioctl(afd, SIOCGIFNETMASK, &afr) == -1) {
             // Error, netmask wasn't found
@@ -567,22 +567,22 @@
             // Return error
             return nil;
         }
-        
+
         // Close the socket
         close(afd);
-        
+
         // Create a char for the netmask
         char *netstring = inet_ntoa(((struct sockaddr_in *)&afr.ifr_addr)->sin_addr);
-        
+
         // Create a string for the netmask
         NSString *Netmask = [NSString stringWithUTF8String:netstring];
-        
+
         // Check to make sure it's not nil
         if (Netmask == nil || Netmask.length <= 0) {
             // Error, netmask not found
             return nil;
         }
-        
+
         // Return successful
         return Netmask;
     }
@@ -599,7 +599,7 @@
         // Set up strings for the IP and Netmask
         NSString *IPAddress = [self wiFiIPAddress];
         NSString *NMAddress = [self wiFiNetmaskAddress];
-        
+
         // Check to make sure they aren't nil
         if (IPAddress == nil || IPAddress.length <= 0) {
             // Error, IP Address can't be nil
@@ -609,41 +609,41 @@
             // Error, NM Address can't be nil
             return nil;
         }
-        
+
         // Check the formatting of the IP and NM Addresses
         NSArray *IPCheck = [IPAddress componentsSeparatedByString:@"."];
         NSArray *NMCheck = [NMAddress componentsSeparatedByString:@"."];
-        
+
         // Make sure the IP and NM Addresses are correct
         if (IPCheck.count != 4 || NMCheck.count != 4) {
             // Incorrect IP Addresses
             return nil;
         }
-        
+
         // Set up the variables
         NSUInteger IP = 0;
         NSUInteger NM = 0;
         NSUInteger CS = 24;
-        
+
         // Make the address based on the other addresses
         for (NSUInteger i = 0; i < 4; i++, CS -= 8) {
             IP |= [[IPCheck objectAtIndex:i] intValue] << CS;
             NM |= [[NMCheck objectAtIndex:i] intValue] << CS;
         }
-        
+
         // Set it equal to the formatted raw addresses
         NSUInteger BA = ~NM | IP;
-        
+
         // Make a string for the address
-        NSString *BroadcastAddress = [NSString stringWithFormat:@"%d.%d.%d.%d", (BA & 0xFF000000) >> 24,
-                                      (BA & 0x00FF0000) >> 16, (BA & 0x0000FF00) >> 8, BA & 0x000000FF];
-        
+        NSString *BroadcastAddress = [NSString stringWithFormat:@"%lu.%lu.%lu.%lu", (unsigned long)(BA & 0xFF000000) >> 24,
+				(unsigned long)(BA & 0x00FF0000) >> 16, (unsigned long)(BA & 0x0000FF00) >> 8, (unsigned long)BA & 0x000000FF];
+
         // Check to make sure the string is valid
         if (BroadcastAddress == nil || BroadcastAddress.length <= 0) {
             // Error, no address
             return nil;
         }
-        
+
         // Return Successful
         return BroadcastAddress;
     }
